Organize code into multiple files.
Split core.py and core_test.py into multiple library and test files.

This enables finer-grained imports (e.g. importing the environment without agents),
which are needed for upcoming work.

Other changes included:
- Change `subject: str` filepath arguments to `subject: str | os.PathLike[str]`.
  - This enables `pathlib.Path` subject filepath arguments.
- Rename `BugLocalizationException` subclassese to end in `Error`.
  - This addresses Google pylint warnings.
